server/backend-python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
google-generativeai==0.3.2
huggingface-hub==0.20.2
requests==2.31.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==41.0.8
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.348
langchain-google-genai==1.0.1
langchain-community==0.0.10
weaviate-client==3.25.3
pinecone-client==2.2.4
qdrant-client==1.7.0
openai==1.6.1
anthropic==0.7.8
google-ai-generativelanguage==0.6.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.25.2
aiofiles==23.2.1
python-socketio==5.10.0
python-socketio[asyncio_client]==5.10.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0